Remove Makefile libvirtd installation target and docker image build on PRs (#28)
* Remove libvirtd installation target from Makefile

libvirtd is installed already on the kind node we use.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>

* Do not build the docker image on PR branches

Dockerfile only installs some dependencies for libvirt-provider
on the kind image. There is no need to build it for each PR once
the image is already there. The docker image doesnt contain any
project specific files in it, only library dependencies.

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>

---------

Signed-off-by: Guvenc Gulce <guevenc.guelce@sap.com>